+++ CHANGELOG.md
@@ -16,9 +16,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added Features and Improvements 🙌:
 - Use predictive back gesture
 - Incorporate user height to calculate BMI-based minimum target weight, replacing the predefined value.
+- Added experimental high contrast mode
 
 ### Bugfix 🐛:
 - Fix system color scheme for monochrome colors, #236
+- Fix using wrong font color in some places
+
 
 ## [0.10.1] - 2025-03-03
 ### Bugfix 🐛: